Add evidence JSON property to issuing a credential
