--- src/pages/Admin/Credentials/views/CredentialManifests/Details/IssueModalWithRender/IssueModalWithRender.scss
@@ -1,64 +1,0 @@
-.dialog {
-    background: var(--grey-5);
-    color: var(--color-white);
-    border: 1px solid var(--grey-10);
-    box-shadow: 0 10px 20px var(--color-black);
-    width: 580px;
-
-    &::backdrop {
-        background: rgba(0,0,0,0.5);
-    }
-
-    &-header {
-        margin-block-end: 1rem;
-        button {
-            .icon-container > svg {
-                width: 24px;
-            }
-        }
-    }
-
-    &-body {
-        margin-inline: 2rem;
-
-        .banner {
-            padding: 1rem;
-            margin-block-end: 1rem;
-            display: flex;
-            gap: 8px;
-
-            &-danger {
-                border: 1px solid var(--color-red);
-                background: var(--red-20);
-                
-                .icon-container {
-                    color: var(--color-red);
-                }
-            }
-
-            &-success {
-                border: 1px solid var(--color-green);
-                background: var(--green-20);
-
-                .icon-container {
-                    color: var(--color-green);
-                }
-            }
-
-            &-warn {
-                border: 1px solid var(--color-orange);
-                background: var(--orange-20);
-
-                .icon-container {
-                    color: var(--color-orange);
-                }
-            }
-        }
-        
-        .button-row {
-            margin-block: 4rem 2rem;
-            justify-content: flex-end;
-        }
-    }
-}
-
