--- src/pages/Admin/Credentials/views/CredentialManifests/Details/IssueModalWithRender/samples/mock.tsx
@@ -1,23 +1,0 @@
-export const credentialInput = {
-    "@context": "string",
-    "data": {
-        "property1": null,
-        "property2": null
-    }
-}
-
-export const schemaInput = {
-    "schema": {
-      "author": "string",
-      "authored": "string",
-      "id": "string",
-      "name": "string",
-      "schema": {
-        "property1": null,
-        "property2": null
-      },
-      "type": "string",
-      "version": "string"
-    },
-    "schemaJwt": "string"
-  }
