+++ src/pages/Admin/Credentials/views/IssuedCredentials/Details/Details.scss
@@ -43,7 +43,7 @@
                 max-width: 100%;
                 box-shadow: 0px 4px 8px rgba(6, 6, 6, 0.5);
                 background: var(--grey-50);
-                background-image: url(https://www.toptal.com/designers/subtlepatterns/uploads/swirl.png);
+                background-image: url('/src/assets/swirl.png');
                 color: var(--color-black);
                 overflow: hidden;
                 padding-block-end: 1.5rem;
