+++ src/pages/Admin/Credentials/views/IssuedCredentials/Modal/Modal.scss
@@ -21,6 +21,11 @@
     &-body {
         margin-inline: 2rem;
 
+        .modal-input-note {
+            font-size: var(--font-xs);
+            margin-block: 0.125rem 0.25rem;
+        }
+
         .banner {
             padding: 1rem;
             margin-block-end: 1rem;
